requirements.txt
# requirements.txt
uvicorn[standard]
fastapi
pydantic
redis
httpx
pillow
celery
pytest
boto3
asgiref
openai
kombu
flower
pytesseract